only delete task if the target is the delete button

+++ app/javascript/packs/deleteTask.js
@@ -1,12 +1,11 @@
 import { $ } from './bling';
 
 const deleteTask = async e => {
+  if (!e.target.classList.toString().match('.task__form--destroy')) return;
   e.preventDefault();
   const { target: { dataset: { taskId, listId } } } = e;
-
   const URL = `${window.location.origin}/lists/${listId}/tasks/${taskId}`;
   const authenticityToken = $('meta[name="csrf-token"]').content;
-
   try {
     fetch(URL, {
       method: 'DELETE',
